TEIID-4883 Excel Translator Filter Empty Column Cells (#951)
* TEIID-4883 Excel Translator Filter Empty Column Cells

Enhance the excel translator to filter column cells that are
empty.  This is only enabled when using a header row and when
the importer.allowEmptyHeaderRow is true.

* Removed unused property

* TEIID-4883 Add flag to skip header row columns will empty value

Added new import property "importer.ignoreEmptyHeaderCells" to 
skip columns in the header that have a empty column name when a header
row is specified.  Also changed the default behavior to stop adding
columns when the first empty header row cell is detected.