JPAMetadataProcessor: Removed addForeignKeys.
It's not clear to me what this is trying to accomplish. The *ToOne
side of a relationship is handled by addSingularAttributes. The
ManyToMany case would surely need a join table created, which this
method does not do. The OneToMany which this method tries to do
something with is unnecessary as far as I can reason, and it results in
an explosion of row results because of joining the relationship in even
the simplist of query (select * from entity).