JPAMetadataProcessor: Avoid PK/FK name collisions.
Old behavior uses a relation's PK as the name of the FK column in the
parent. If the entities happen to use the same name for their IDs you
get a collision and only one column ends up exposed. I for one always
name my entity ids 'id'.

This change instead names the FK column(s)
$parentAttribute_$childIdProperty.