TEIID-5798 adding the concept of a row policy
this is roughly based on pg, but the name is scoped to both the resource
and the role - that's why drop takes both arguments.

it has been simplified to only specify a using condition, which defaults
to a constraint/check.  if you want to create separate read/write
policies, you need add two policies instead of one.

this does not yet attempt to rewrite the old permission condition logic
as policies