TEIID-4594 generalizing the filtering
these changes remove the needs for partial filtering.  all comparisons
will be handled at the translator level or throw an exception if
invalid.

there are now three ways predicates are applied - in the path, against
the path only before the file is processed, and as a row filter

this requires the use of a userdefined predicate / filter visitor